requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\Ryan\discord-bloom\bloom.py: 6
aioschedule == 0.5.2

# C:\Users\Ryan\discord-bloom\bloom.py: 1,2,3
discord.py == 1.6.0

# C:\Users\Ryan\discord-bloom\bloom.py: 10
emoji == 0.6.0
